Fix issues on encoding (#646)
- Fix the issue where the default encoder was ignored.
- Fix the issue where the default encoder type was always video.
- Also get rid of unnecessary copy

+++ src/libspdl/core/conversion.h
@@ -39,5 +39,5 @@ VideoFramesPtr convert_rgb_array(
     size_t height,
     size_t width,
     Rational time_base,
-    int pts);
+    int64_t pts);
 } // namespace spdl::core
